scraper/requirements.txt
playwright==1.43.0
supabase==2.4.3
python-dotenv==1.0.1
requests==2.31.0
openai==1.25.0
python-dateutil==2.9.0
